fix : BD-213 PR 코멘트 내용을 반영하여 날짜 계산 로직의 에러를 처리합니다.
